requirements.txt
beautifulsoup4==4.12.3
biopython==1.84
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
# Additions for embedding and retrieval
chromadb
sentence-transformers
transformers
torch
